requirements.txt
environs==11.0.0
python-dotenv==1.0.1
aiofiles==24.1.0
ConfigArgParse==1.7